+++ html/signin.html
@@ -41,6 +41,6 @@
             </div>
         </div>
     </div>
-    <script type="module" src="/script/sign.js"></script>
+    <script type="module" src="/script/signin.js"></script>
 </body>
 </html>
