+++ script/sign.js
@@ -2,14 +2,13 @@ const $emailErrorMessage = document.querySelector('.email_error_message');
 const $pwdErrorMessage = document.querySelector('.pwd_error_message');
 const $email = document.querySelector('.email_input');
 const $pwd = document.querySelector('.pwd_input');
-const $submit = document.querySelector('.form__sign');
 const $pwdEye = document.querySelectorAll('.password-eye');
 
 let emailValid = false;
 let pwdValid = false;
 
 const REGEMAIL = /^[A-Za-z0-9\-]+@[A-Za-z0-9]+\.[a-z]/;
-function emailErrorMessage(e){
+function emailErrorMessage(){
     emailValid = false;
     if($email.value === ""){
         $emailErrorMessage.textContent = "이메일을 입력해주세요."
@@ -30,7 +29,7 @@ function emailErrorMessage(e){
 $email.addEventListener("focusout",emailErrorMessage);
 
 
-function pwdErrorMessage(e){
+function pwdErrorMessage(){
     pwdValid=false;
     if($pwd.value === ""){
         $pwdErrorMessage.textContent = "비밀번호를 입력해주세요."
@@ -45,25 +44,6 @@ function pwdErrorMessage(e){
 }
 $pwd.addEventListener("focusout",pwdErrorMessage);
 
-function signinValidCheck(e){
-    if(emailValid && pwdValid){
-        return;
-    }
-    if(!emailValid){
-        $emailErrorMessage.textContent = "이메일을 확인해주세요."
-        e.preventDefault();
-    }
-    if(!pwdValid){
-        $pwdErrorMessage.textContent = "비밀번호를 확인해주세요."
-        $pwdErrorMessage.style.display ="block";
-        e.preventDefault();
-    }
-}
-$submit.addEventListener('submit',signinValidCheck);
-$submit.addEventListener('submit',pwdErrorMessage);
-$submit.addEventListener('submit',emailErrorMessage);
-
-
 function pwdEyeOnOff(e){
     let eyeOn = e.target.src.includes('eye-on');
     if(eyeOn){
@@ -77,4 +57,4 @@ function pwdEyeOnOff(e){
 }
 $pwdEye[0].addEventListener('click',pwdEyeOnOff);
 
-export {$email,$emailErrorMessage,$pwdErrorMessage,$pwd,pwdEyeOnOff,$pwdEye,emailValid,emailErrorMessage}
+export {$email,$emailErrorMessage,$pwdErrorMessage,$pwd,pwdEyeOnOff,$pwdEye,emailValid,pwdValid,emailErrorMessage,pwdErrorMessage}
